feat: switch to using vllm
* when cuda is not available, we are still using llama.cpp as a fallback
