+++ pyproject.toml
@@ -22,16 +22,16 @@ pytest-mock = "3.12.0"
 aiohttp = "3.9.5"
 async-lru = "2.0.4"
 boto3 = "^1.28.56"
-fastapi = "0.109"
+fastapi = "0.115.0"
 fastapi-versionizer = "3.0.4"
 faster-whisper = "1.0.3"
-prometheus-client = "0.19.0"
-prometheus-fastapi-instrumentator = "6.1.0"
+prometheus-client = "0.20.0"
+prometheus-fastapi-instrumentator = "7.0.0"
 pyjwt = {extras = ["crypto"], version = "^2.8.0"}
 python = "~3.11"
 redis = "5.0.1"
-torch = ">=2.0.0,<2.1.0"
-torchaudio = ">=2.0.1,<2.1.0"
+torch = "2.4.0"
+torchaudio = "2.4.0"
 uvicorn = {extras = ["standard"], version = "0.29.0"}
 uuid6 = "^2024.1.12"
 pyyaml = "^6.0.1"
@@ -41,6 +41,7 @@ langchain = "^0.3.0"
 langchain-openai = "^0.2.0"
 av = "^12.3.0"
 pybase64 = "^1.4.0"
+vllm = "^0.6.2"
 
 [build-system]
 build-backend = "poetry.core.masonry.api"
