+++ run.sh
@@ -1,7 +1,13 @@
 #!/bin/sh
-cd llama.cpp
-make llama-server
-cd ..
 
-export LLAMA_N_CTX=32000
+if nvcc --version
+then
+  export CUDA_VISIBLE_DEVICES=0
+else
+  cd llama.cpp
+  make llama-server
+  cd ..
+fi
+
+export LLAMA_N_CTX=8192
 poetry run python -m uvicorn skynet.main:app --reload
