+++ skynet/modules/ttt/openai_api/app.py
@@ -2,18 +2,21 @@ import subprocess
 
 from skynet import http_client
 from skynet.env import (
+    llama_cpp_server_path,
     llama_n_batch,
     llama_n_ctx,
     llama_n_gpu_layers,
     llama_path,
     openai_api_base_url,
-    openai_api_server_path,
     openai_api_server_port,
+    vllm_server_path,
 )
 from skynet.logs import get_logger
 from skynet.modules.monitoring import OPENAI_API_RESTART_COUNTER
+from skynet.utils import get_device
 
 proc = None
+use_vllm = get_device() == 'cuda'
 
 log = get_logger(__name__)
 
@@ -23,16 +26,28 @@ def initialize():
 
     global proc
 
-    proc = subprocess.Popen(
-        f'{openai_api_server_path} \
-            --batch-size {llama_n_batch} \
-            --ctx-size {llama_n_ctx} \
-            --flash-attn \
-            --model {llama_path} \
-            --n-gpu-layers {llama_n_gpu_layers} \
-            --port {openai_api_server_port}'.split(),
-        shell=False,
-    )
+    if use_vllm:
+        openai_api_server_path = vllm_server_path
+        proc = subprocess.Popen(
+            f'python -m {openai_api_server_path} \
+                --model {llama_path} \
+                --gpu_memory_utilization 0.95 \
+                --max-model-len {llama_n_ctx} \
+                --port {openai_api_server_port}'.split(),
+            shell=False,
+        )
+    else:
+        openai_api_server_path = llama_cpp_server_path
+        proc = subprocess.Popen(
+            f'{openai_api_server_path} \
+                --batch-size {llama_n_batch} \
+                --ctx-size {llama_n_ctx} \
+                --flash-attn \
+                --model {llama_path} \
+                --n-gpu-layers {llama_n_gpu_layers} \
+                --port {openai_api_server_port}'.split(),
+            shell=False,
+        )
 
     if proc.poll() is not None:
         log.error(f'Failed to start OpenAI API server from {openai_api_server_path}')
@@ -41,14 +56,10 @@ def initialize():
 
 
 async def is_ready():
-    # https://github.com/ggerganov/llama.cpp/blob/master/examples/server/README.md#get-health-returns-heath-check-result
     try:
-        response = await http_client.get(f'{openai_api_base_url}/health')
+        await http_client.get(f'{openai_api_base_url}/health', 'text' if use_vllm else 'json')
 
-        if response.get('error'):
-            return False
-
-        return response.get('status') == 'ok'
+        return True
     except Exception:
         return False
